[qemu,opentitanlib] Remove warning for reading floating GPIO
As per the comment, it is a common pattern that means that unfortunately
we do this several times with the SPI console. Ideally in the future
better synchronization mechanisms should be used such that the SPI
console TX pin is not read from until it is pulled up, but for now we
reduce QEMU test noise by silencing this warning.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>