[qemu,opentitanlib] Fix GPIO PTY reads
This was using a `BufReader` to buffer the incoming PTY contents but was
directly checking the underlying PTY internal length to determine if
there was data left to read, which meant that if there were multiple
lines pending, the first read would buffer subsequent lines but not read
them out. This led to a delay in GPIO handling where only a single GPIO
event would be handled even if multiple were received, though they would
still all be handled in order.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>