[qemu,opentitanlib] Handle GPIO input forwarding
QEMU's GPIO protocol also has an undocumented command to forward
the GPIO inputs that it reads to the backend / remote host for
debugging, to track and test the actual GPIO input value.

Although we do not use this currently this commit opts to store it as we
may want to use it for testing purposes in the future.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>