[p256/p384/cryptolib] Protect scalar in ecdh
The Fi simulation from #28549
found two vulnerabilities from instruction skips in ecdh p256 and two in p384.
These stem from offsetting pointers in the hardened_memcpy.
To protect against them a hardened_memeq is added.

In addition, stack operations entering into hardened_memcpy can leave
for a vulnerability where the destination or source is not loaded (and
thus not set back after stack operations). To add redundancy, we load
the dest and src twice in the hardened_memcpy function.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>