[ottf,spi_console] Use read addr instead of CS for sync with TX ready pin
The current OTTF SPI console implementation, when using a TX ready GPIO
indicator pin, watches to see the chip select signal (the SPI Device's
`STATUS.csb`) toggle 4 times corresponding to asserting/deasserting for
the header read and then the payload read. Using this, the device can
determine when the host has started/finished reading and synchronize
accordingly.

This is not ideal because it depends on a direct signal (CSb) which (a)
could be problematic for different clock speeds if the CS is asserted
for a short time and Ibex does not poll fast enough, since there is no
interrupt here, and (b) is very difficult for emulation flows to manage,
such as QEMU.

Instead, since the SPI device exposes the last read address (and the
exact amount of data to be read is known ahead of time), we can just
check the last read address to determine when the host is finished
reading. We still deassert the TX ready pin after either seeing CS go
low or a read address change, but we now know the transaction is
finished when CS is high and we see a match on our target address.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>