[qemu,opentitanlib] Handle large SPI transactions
Because the QEMU SPI protocol allows the construction of frames with up
to 65535 byte payloads, it is possible to send SPI transmissions that
exceed the standard kernel TTY buffer length of 4 KiB. Although we claim
the SPI supports bidirectional full-duplex transport, in reality we are
interfacing with the transfer protocol in a half-duplex fashion.

To retain performance by minimizing PTY writes/reads whilst also
handling these cases, chunk incoming transfers based on this TTY buffer
size.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>