[qemu,opentitanlib] Increase SPI Serial Port timeout
Increase the timeout on the SPI Serial port from 1 second to 5 seconds.
This is needed because, for very large SPI transfers, the SPI flash
buffer boundary can be crossed (a 'flipbuf' event for the SPI device).
In such a scenario, the SPI device hardware generates an interrupt for
SW to handle and potentially modify the buffers appropriately. To allow
for this, it emulates a large delay which is cancelled by SW clearing
the interrupt.

If SW chooses not to, then the full delay is emulated while handling the
SPI transaction. If this delay is configured high enough (and we use a
sizable icount shift), we can sometimes see these large SPI transactions
exceeding 1 second by a small amount. While this delay should likely be
reduced in QEMU emulation, there's no harm here in increasing the
timeout to be able to handle different emulation behaviour.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>