[otp_ctrl,dv] Pass some constructor arguments more explicitly
This is motivated by the fact that I want to get rid of the middle
optional argument for bit_toggle_cg_wrap. There is only non-trivial
use of the argument in the repository and it is the creation of
status_csr_cg in this module.

Inline the calculation here instead, and name the toggle_cov_en
argument, which lets us not explicitly provide the middle argument.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>