[alert,dv] Simplify how we constrain ping_delay
This number has been very confusing! At the start of 2024, functional
simulations weren't sending pings at all. When I made change
a771270fab3 that year, I interpreted ping_delay as "the number of
cycles to wait before sending a ping". So that we could send
back-to-back ping items but have reasonable gaps between pings, I was
configuring this to be large (> 10,000).

What I didn't notice is that some sequences constrained it to be more
like 10! I think the original idea was that a virtual sequence would
decide to send a ping (for some reason), which would then take a
handful of extra cycles before it was sent.

Unfortunately, this was meaning that some virtual sequences were
sending *lots* of ping transactions, with gaps of just 10 cycles. The
change in this commit switches everything across to the new meaning.
We now send back-to-back items, which translate into periodic ping
requests.

There are a couple of extra tidy-ups in the commit

 - Get rid of use_seq_item_ping_delay. I accidentally removed the code
   that was reading this in commit 1eb96b30. The new behaviour is
   probably better, so let's just use it.

 - Tweak a wait in entropy_src_base_vseq. This was added in commit
   a3f91eec28e, which wanted to make sure the sequence waited long
   enough to see alerts that would be triggered by the previous
   register write. With the new meaning of ping_delay, we definitely
   don't need to wait for it.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>