Make PLIC interrupt 0 (no interrupt) more robust
In PLIC, interrupt ID 0 is special - it is reserved for "no interrupt". The way this is implemented in `rv_plic` is by simply asking nicely for integrators to tie the interrupt ID 0 input to 0.

This is suboptimal for a few reasons:

1. It requires adding a dangerous `assume` for FPV - this is directly added inside `rv_plic.sv`. Unless there's some flow magic to convert this to an `assert` (I didn't see anything) then this can cause nonsensical formal results if you accidentally violate that assumption.
2. It allows the possibility of integration mistakes causing an interrupt 0.

This change instead just ignores that input and ties it to 0 internally. I would have liked to remove it entirely so there is no chance of making that mistake, but that is a much bigger change that requires changing the external interface.

Signed-off-by: Tim Hutt <tim.hutt@scisemi.com>