[opentitanlib,qemu] Add additional SPI console TX sync for QEMU
This is only enabled if the QEMU interface is configured for the SPI
console - any other interface will retain the existing behaviour and
avoid the additional poling from the host side.

For QEMU specifically, we introduce a host-side poll to see the TX ready
pin go low after reading out the header (but before reading out the
payload). This ensures we don't run into synchronization issues due to
emulation for very small SPI transfers in which the entirity of the
header & payload are read out by the host before the device has a chance
to run enough guest SW to signal that the GPIO ready pin should go low.

In such a scenario, the device would look to host SW as if it is ready
to send more data, even though in reality the guest SW has barely
progressed enough to register that the host has started reading, causing
the two devices to get out of sync.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>