[bazel,qemu] Enable SPI console with TX indicator test for QEMU
Enables the `ottf_console_with_gpio_tx_indicator` test target for the
QEMU execution environment, as this test exercises functionality that is
required to be supported for provisioning flows.

As QEMU does not currently implement the relevant Pinmux connections
between OpenTitan and GPIO, a minor extension is made to the test to
allow users to optionally specify which GPIO pin should be used as the
ready-indicator, rather than hardcoding to IOA5. This is then overriden
for the QEMU execution environment only.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>