[opentitanlib] Split UsbBackend into a Context and a Device
The name of the type itself does not make sense: UsbBackend has
associated methods which truly act like a context, but the struct
itself corresponds to a device.

This commit properly splits this into a context and a device.
Even though the context is an empty struct, this is done for
consistency with other potential backends which may require state.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>