[opentitanlib] Introduce an abstraction over the USB backend
Currently there is a USB backend struct which wraps around rusb
and provides some utility functions. In the future, we may want
some transports to implement USB in a different way (e.g. QEMU
implements USB over a virtual channel). To do, the first step
is to create an abstraction over the existing `rusb` implementation.

This PR introduces a trait which is then implemented by the rusb
backend.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>