[opentitanlib] move `set_break` from `ConsoleDevice` to `Uart`
Previously `set_break` needs to be on `ConsoleDevice` because `UartConsole`
expects it to be there, so it can turn stdin breaks into console breaks.
This feature is now moved to `opentitantool console` (which was the only place
this feature was used), so the `set_break` doesn't need to be on `ConsoleDevice`
anymore (and also this capability is only available on `Uart` impls anyway).

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>