[memory/cryptolib] Launder status output
Since the hardened memory operations just return OTCRYPTO_OK the
compiler optimizes the status output which circumvents the control flow
integrity. An example is

20011276:             |     |         9e2080e7                  jalr
		      -1566(ra) # 2002bc54 <hardened_memcpy>

2001127a:             |     |         73900513                  li
a0,1849

Launder the status output to circumvent this. Once there is another call
in place which makes the possibility to return a different status, this
launder is no longer needed.

The assembly files were checked for similar patterns, only a single call
in RSA had a similar result.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>