[opentitanlib] Implement stub USB context for QEMU
This commit makes the QEMU backend able to create a USB context on
demande. For now the context is mostly stubbed: when requested, the
USB minihost will be launched so that the device will be enumerated,
but not `UsbDevice` can be created by the context. This is still
useful for some tests that only require the device to be enumerate/
configured and do not require interaction over USB with the harness.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>