[xbar,rtl] support fifo_depth>1
This commit adds the possibility to increase the fifo depth in the xbar
to values > 1 to support multiple outstanding transactions.

Why this is beneficial:
The ibex instruction cache issues two 32b requests to the flash controller.
Inside xbar_main a pipeline register is added to break the critical path
to the flash. The pipeline register is added with a fifo of depth=1 for
req and rsp data and effectively inserts a bubble after each request and
response because the fifo is immediately full. Ibex and flash_ctrl can
deal with up to 2 outstanding transactions.

The impact on performance is low because the instruction cache reads the
critical word first and hides the additional latency that is inserted by
the fifo with depth=1. Nonetheless, in phases with many cache misses,
the performance can be improved at the price of an additional fifo entry.

Signed-off-by: Michael Gautschi <mgautschi@lowrisc.org>