[pentest] Add local alerts
Add the interface to enable local alerts and read out the configuration
of the local alerts. After each fault injection test, print back the
triggered local alerts.
The local alerts can be enabled or disabled by the user. They are always
set to class A and they are always fatal.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>