[python] Add pypi dvsim 1.6.1 to project python environment
This allows the in-tree dvsim (util/dvsim/dvsim.py) to be used alongside the
external dvsim (github.com/lowRISC/dvsim) during a transition period.
The existing dvsim codebase will continue to function exactly as before, and is
not affected in any way by adding the new package into the environment.

The replacement dvsim tool is available as an unprefixed binary on the path once
installed. This should not conflict with any existing scripting, which all uses
the existing dvsim entrypoint script via relative path from the project root.

[old dvsim] $ util/dvsim/dvsim.py
[new dvsim] $ dvsim

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>