[otbn,sim] Add otbnsim command to stall for one cycle instead of retiring
This adds the possibilities to let the model stall for a cycle instead of retiring.
This is a workaround to model the ld_insn injection error from the otbn_ctrl_redun_vseq.sv
test.

When a load instruction is injected in parallel to the actual instruction, the OTBN will
not execute the current instruction and will stall (due to how the ld_insn signal factors
into the stall logic). Modeling this behaviour is complicated and would require to duplicate RTL
implementation details in the model. Thus, we add this command/flag which UVM can use to
signal that we should enter the stall state.

Signed-off-by: Pascal Etterli <pascal.etterli@lowrisc.org>