[otbn,sim] Extend DMEM model with valid flag for each entry
Previously the DMEM integrity validity was modelled with an Optional[Int] type
where an int value represented valid and None an invalid state.
The DMEM integrity test then did simply invalidate the whole DMEM by setting
all parts to None.

This is incompatible with a delayed escalation as a failing instruction (e.g.,
a LW) will trigger the escalation but will still retire. As of this it actually
must load a value and a None value cannot be written to a register. To solve
this, we model the integrity with a dedicated valid flag and return the current
value as well. The test now only has to specify that a comparison mismatch is
tolerated.

It would also be possible to modify the insn implementation that if None is
returned a fixed value is written to the register. However, in this case the
insn code would not exactly represent the HW behaviour. As this insn code is
converted to pseudo code for this the ISA description, this approach is less
favourable. Keeping the value field also allows to extend the DMEM model in future
to fully model the injected error (i.e., extending the otbnsim with a command
which lets set DMEM content to the faulted values). With this the "tolerate
mismatch" workaround could be fixed properly.

Signed-off-by: Pascal Etterli <pascal.etterli@lowrisc.org>