[manuf] Remove manual GPIO toggles on final hash transmission
When sending the final hash of all personalization objects to the host
to confirm the integrity of provisioning, the current implementation
manually toggles the GPIO TX Ready indicator used with the SPI console
for the transmission to high before the transmission, and low after.

In reality, the ujson serialization of the hash will call through to the
OTTF SPI console via `base_print` which will handle all of the logic for
toggling the pin for you when the data is ready in the SPI read buffer.
The current logic in fact makes timing assumptions: it assumes that the
data will be ready in the SPI read buffer by the time the host tries to
read the frame header based on the GPIO pin going high . This works now
because silicon & FPGAs are fast enough, but might not be true if the
SPI console becomes more complicated / slower, or at different clock
speeds, or in sim/emulation environments (e.g. QEMU, which does not
synchronize I/O and guest SW well, sometimes errors on this case).

The fix is simple: let the SPI console handle the GPIO pin for us, and
stop trying to manually handle it ourselves.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>