[sram_ctrl,doc] Update `EXEC` register description to note HW Debug
The existing sram_ctrl documentation for the `EXEC.EN` register field
is not incorrect, but is somewhat misleading by omission as it does
not reflect the alternate hardware mechanism for enabling SRAM
execution via the lc_ctrl's HW debug enable signal. This is part of
the sram_ctrl.sv RTL so should probably be included, so add it here.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>