[otbn,dv] Isolate HW loop stack countermeasure testing from integrity checks
When testing the countermeasures for the OTBN hardware loop stack, the current_loop_valid
can be driven to high as the loop stack counter is forced to zero. This fault should be
detected and assert the `hw_err_o` of the loop controller. As in this test no SW is run the
loop stack has undefined values (x). When the `current_loop_valid` is high, the integrity
check of the loop address (current_loop_intg_err) is OR factored into the `hw_err_o` as well
leading to undefined error bits. We prevent this propagation by forcing the
`current_loop_intg_err` to zero / we ignore it. This ensures that the countermeasures must
detect the fault. This is ok because in real HW the integrity error is OR combined and
therefore only additionally drives the error signal (X|1 = 1).

See also #28888

Signed-off-by: Pascal Etterli <pascal.etterli@lowrisc.org>