[cryptotest] Fix private key `keyblob_length`
The ecdsa cryptotest used `sizeof(private_key_masked)` as the
`keyblob_length`. However, this is not correct as this sizeof
contains the entire size of the private key, which includes the
checksum. However, in `keyblob_length` we just expect the keyblob
length, with is the length of both key shares.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>