[ci] Use BLOCKFILE of PR target branch instead of the default branch
See https://github.blog/changelog/2025-11-07-actions-pull_request_target-and-environment-branch-protections-changes/
After these changes to Github Actions took effect, all PRs to any branch
of OpenTitan changed to use the `master` workflow for the target branch
protection. While this was fine for calculating the PR diff (using the
GH CLI), it also meant that the BLOCKFILE being used by the check script
was the BLOCKFILE on master.

This is not ideal, as we want each non-default branch to be able to
specify a BLOCKFILE for files that should be blocked specifically on
that branch. For example, on `earlgrey_1.0.0`, we do not want to permit
any further RTL changes.

This change ensures that the PR target branch's BLOCKFILE is used to
provide appropriate protections.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>