[pentest] Move alert handling cryptolib to wrapper
In the pentest calling of the cryptolib we use TRY functions that return
early if a bad status is detected. However, if we return early the
alerts are left unset. In order to prevent this, we set the alerts in
the wrapper logic instead.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>
(cherry picked from commit 7035b02011c798e5f6dd45ebda265fe7be714933)