[otlib,usb] Make device's serial number optional
The USB specification does not mandate that a device must have a serial
number so there is no reason to hardcoce this assumption in the USB
abstraction. Some backends like hyperdebug always interface with devices
with a serial number, in this case it is correct to require it.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 2a88b7f735a31aa11fa2699227c8e5651dd8739e)
(with nontrivial modifications)