[pentes] Set pacing higher
We found that the python_tests in earlgrey_1.0.0 were flaky causing json
errors (no output). Bisecting through commits did not reveal a change
causing the flaky result and testing reveals master is stable, but
earlgrey_1.0.0 is not, and that all python_tests were flaky but not the
rust tests.
After additional debugging, doubling the pacing value seems to clear
this up.

Unsure why this was in eg100 but not in master, but I am copying the
change in both branches.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>
(cherry picked from commit 672d5bd4e1754b672701466fe493629d71e44e6f)