[adc_ctrl,dv] Simplify adc_ctrl_filters_polled_vseq
This was a bit complicated: all it's really doing is to ask a couple
of existing sequences to set some variables in a series.

This patch simplifes things dramatically. For a start, it runs one
vseq per channel, which means the new adc_ctrl_channel_ramp_vseq can
be a bit simpler than adc_ctrl_random_ramp_vseq was.

It also avoids messing around with the "rising" flag: we can just use
a start and end value and have the direction implicit. Indeed, it
could have been even shorter(!), but I decided it was wiser not to rely
on unsigned wrapping in adc_ctrl_channel_ramp_vseq::body.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>