[AES/pre_syn] Add a yosys synthesis bazel target
In the pre_syn folder from the AES there are bash scripts to generate
area and timing reports together with a netlist for the AES core using
OpenSTA and Yosys with the nangate45 library.
We integrate these scripts using a Bazel target such that the reports
can be generated using dependencies all within Bazel.

This default Bazel target does not yet pull in OpenSTA, hence timing
results are not yet present but it generates the area reports and the
netlist.

Note that we add the nangate45 lib file in the AES repo.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>