[AES/prolead] AES prolead nightly target
Add a target to evaluate the aes_cipher_core in prolead with a nightly
CI target. The test costs 100s and is adapted to have an extremely low
probability for a false positive.
This way, the AES RTL is evaluated whether the masking is working as
intended or whether a critical failure was introduced.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>